+++ Cargo.toml
@@ -17,3 +17,4 @@ serde_json = "1.0"
 env_logger = "0.11"
 tokio = { version = "1.0", features = ["full"] }
 chrono = { version = "0.4", features = ["serde"] }
+axum = "0.7"
